mips: Change the integer and fp register widths to be 64 bits.
This makes them more consistent with the other architectures, helping
to facilitate having a single, unified type across ISAs in the future.

Reviewed-on: https://gem5-review.googlesource.com/c/13620
Maintainer: Jason Lowe-Power <jason@lowepower.com>

+++ src/arch/mips/registers.hh
@@ -278,11 +278,11 @@ const int NumMiscRegs = MISCREG_NUMREGS;
 
 const int TotalNumRegs = NumIntRegs + NumFloatRegs + NumMiscRegs;
 
-typedef uint32_t IntReg;
+typedef uint64_t IntReg;
 
 // floating point register file entry type
-typedef uint32_t FloatRegBits;
-typedef float FloatReg;
+typedef uint64_t FloatRegBits;
+typedef double FloatReg;
 
 // cop-0/cop-1 system control register
 typedef uint64_t MiscReg;
